+++ src/CheckIn.js
@@ -9,6 +9,8 @@ import {
   intlShape
 } from 'react-intl';
 import moment from 'moment-timezone';
+import createInactivityTimer from 'inactivity-timer';
+
 import {
   Paneset,
   Pane,
@@ -34,7 +36,10 @@ import { IfPermission } from '@folio/stripes/core';
 import PrintButton from './components/PrintButton';
 import FeesFinesOwnedStatus from './components/FeesFinesOwnedStatus';
 import FeeFineDetailsButton from './components/FeeFineDetailsButton';
-import { convertToSlipData } from './util';
+import {
+  convertToSlipData,
+  getCheckinSettings,
+} from './util';
 import styles from './checkin.css';
 
 class CheckIn extends React.Component {
@@ -51,7 +56,11 @@ class CheckIn extends React.Component {
     }),
     onSessionEnd: PropTypes.func,
     change: PropTypes.func,
-    resources: PropTypes.object,
+    resources: PropTypes.shape({
+      checkinSettings: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
+    }),
     mutator: PropTypes.shape({
       query: PropTypes.shape({
         update: PropTypes.func,
@@ -69,14 +78,47 @@ class CheckIn extends React.Component {
     this.showInfo = this.showInfo.bind(this);
     this.renderActions = this.renderActions.bind(this);
     this.handleOptionsChange = this.handleOptionsChange.bind(this);
+    this.timer = undefined;
 
     this.state = {
       showPickers: false
     };
   }
 
-  componentDidMount() {
+  componentDidUpdate() {
+    const {
+      resources: {
+        checkinSettings,
+        scannedItems,
+      },
+      onSessionEnd,
+    } = this.props;
+
     this.focusInput();
+
+    if (this.timer) return;
+    if (!checkinSettings || !checkinSettings.records || checkinSettings.records.length === 0) return;
+
+    const parsed = getCheckinSettings(checkinSettings.records);
+    
+    if (!parsed.checkoutTimeout) {
+      this.timer = null;
+      return;
+    }
+
+    if (scannedItems.length) {
+      this.timer = createInactivityTimer(`${parsed.checkoutTimeoutDuration}m`, () => {
+        onSessionEnd();
+      });
+
+      ['keydown', 'mousedown'].forEach((event) => {
+        document.addEventListener(event, () => {
+          if (this.timer) {
+            this.timer.signal();
+          }
+        });
+      });
+    }
   }
 
   focusInput() {