+++ src/Scan.js
@@ -16,7 +16,6 @@ import {
   isEmpty,
   keyBy,
   upperFirst,
-  uniqBy,
 } from 'lodash';
 
 import {
@@ -60,7 +59,10 @@ class Scan extends React.Component {
       }),
       items: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
-      })
+      }),
+      checkinSettings: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
     }),
 
     mutator: PropTypes.shape({
@@ -151,6 +153,11 @@ class Scan extends React.Component {
       type: 'okapi',
       path: 'circulation/end-patron-action-session',
       fetch: false,
+    },
+    checkinSettings: {
+      type: 'okapi',
+      records: 'configs',
+      path: 'configurations/entries?query=(module=CHECKOUT and configName=other_settings)',
     },
   });